% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplotter.R
\name{corrplotter}
\alias{corrplotter}
\title{Function to produce correlation panel plots}
\usage{
corrplotter(dataframe, colIDs, hist.breaks, hist.col, hist.names, text.cex,
  neg.color, mid.color, pos.color, par.oma, par.tcl, par.mgp, par.lab,
  left.text, bottom.text, expansion, ...)
}
\arguments{
\item{dataframe}{The object you want to pull data from to plot.}

\item{colIDs}{Optional vector of column IDs or names you want to plot from dataframe.}

\item{hist.breaks}{The number of breaks desired in the diagonal histograms. Default is
10.}

\item{hist.col}{The color of the histograms. Default is black.}

\item{hist.names}{The titles to be printed above histograms, if desired. Will be left
blank if not provided.}

\item{text.cex}{The size of the summary stat text in the upper half of the plots.
Default is 2.}

\item{neg.color}{The background color of the panel in the upper half of the plots for
correlations with negative coefficients. It is very important that this be formatted as
a vector of length four, minimum 0, maximum 1, corresponding to rgb + alpha channel.
All other formats will fail. Default is red, with alpha 0.5.}

\item{mid.color}{The background color of the panel in the upper half of the plots for
non-existent correlations. Format the same as neg.color. Default is white.}

\item{pos.color}{The background color of the panel in the upper half of the plots for
correlations with positive coefficients. Format the same as neg.color. Default is blue,
with alpha 0.5.}

\item{par.oma}{Used to format the space around the entire plot (the area outside all
panels). The format is c(bottom, left, top, right). Default is c(4, 4, 2, 2).}

\item{par.tcl}{Used to control the length of the tick marks inside the panels.}

\item{par.mgp}{Used to control the distance of the labels from the tick marks inside
the panels. Default is c(2, 0.2, 0).}

\item{par.lab}{Used to control the number of tick marks inside the panels. Format is
c(x, y, len). Although len is not implemented in R, it is necessary to pass along.
Default is c(5, 5, 7).}

\item{left.text}{The labels to be printed along each panel on the left margin of the
figure. Must be the same length as the number of columns to be printed. If you do not
want to print text along the last panel (the histogram), simply make the last element
of the vector be equal to "".}

\item{bottom.text}{Like left.text, but for the bottom axis labels.}

\item{expansion}{Controls how much the histogram is shrunk down to facilitate plotting
of histogram names. Also controls how zoomed out the scatterplots are, which might be
useful for ensuring dots do not interfere with axis labels.}

\item{...}{Other arguments that can be passed to plot, including cex for the size of
the dots in the scatterplot.}
}
\value{
A paneled correlation plot.
}
\description{
Fairly versatile function to produce paneled correlation plots, where the lower
triangle of panels shows the x,y relationship, the diagonal shows a histogram of each,
column of data, and the upper triangle summarizes the correlation.
}
\details{
This function sends out jobs to as many cores as are specified. Each 
randomizes the input CDM according to all defined null models, then calculates each
observed metric on each randomized matrix.
}
\examples{
exData <- data.frame(
col1=jitter(sort(rnorm(n=200, mean=0, sd=1), decreasing=TRUE), amount=1),
col2=jitter(sort(rnorm(n=200, mean=10, sd=1)), amount=4),
col3=jitter(sort(rnorm(n=200, mean=-3, sd=1)), amount=3))

corrplotter(dataframe=exData, hist.breaks=20, hist.col="purple",
hist.names=c("one","two","three"), left.text=c("units1", "units2", ""),
bottom.text=c("other_units1", "other_units2", ""),
text.cex=2, neg.color=c(1,0,0,1), mid.color=c(1,1,0,1), pos.color=c(0,1,0,1),
expansion=0.5, cex=0.5, col="gray")
}

